Several changes made as per code-review.
